Add some state and locking to server view
Added current stats, selected line position, number of lines and a mutex
to protect it all. Had to tweak config parsing a bit to return a slice
of pointers, otherwise go s.monitor() would run on the same copy of
server.
Also added cursor movement method. So far not used.